backend/requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.0
h11==0.13.0
idna==3.4
motor==3.0.0
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
stripe==4.2.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
